Refactor geofencing logic and clean up legacy components
Removed legacy geofencing UI and logic from GeofenceToggle and GeofencingExample, replacing them with placeholders. Refactored useGeofencing hook to a minimal interface. Rewrote GeofencingService to a minimal version and added a legacy copy for reference. Updated AlertsScreen to handle geofence management directly and improved geofence add/remove logic. Cleaned up related Redux middleware, saga, and slice to align with the new geofencing approach.
